atomic_energy/examples.py
Absolutely, I understand the need for a more elaborate example. The approach involves Django signals to trigger a Celery task for asynchronous handling of file transfers to a remote media host.

Here's a more detailed example:

    models.py remains the same:

python

from django.db import models

class Media(models.Model):
    file = models.FileField(upload_to='media')

    signals.py:

python

from django.db.models.signals import post_save
from django.dispatch import receiver
from myapp.tasks import transfer_media_to_host  # Replace with your actual task

@receiver(post_save, sender=Media)
def send_media_to_host(sender, instance, created, **kwargs):
    if created:  # Only trigger on creation, not updates
        transfer_media_to_host.delay(instance.file.url)  # Trigger Celery task

    tasks.py (Celery task):

python

from celery import shared_task
import requests

@shared_task
def transfer_media_to_host(file_url):
    media_host_url = 'http://your_media_host.com/upload'  # Replace with your media host URL
    try:
        # Logic to send file to media host (example using requests)
        response = requests.post(media_host_url, files={'file': open(file_url, 'rb')})
        if response.status_code == 200:
            print('File transferred to media host successfully!')
        else:
            print('File transfer to media host failed:', response.status_code)
    except Exception as e:
        print('Error transferring file:', e)

This setup utilizes Django signals to trigger the transfer_media_to_host Celery task whenever a new Media object is created. The task asynchronously handles the transfer of the media file to a remote media host using the requests library. Adjust the URLs and logic within the task to match your specific use case and API requirements.

This design keeps the file transfer process asynchronous, avoiding delays in the application's response time. Adjust the logic within the task as needed to fit the authentication, file transfer method, and error handling required for your media host integration.