requirements.txt
amqp==5.2.0
asgiref==3.7.2
astroid==2.15.5
async-timeout==4.0.3
bandit==1.7.5
bcrypt==4.1.2
beautifulsoup4==4.12.2
billiard==4.2.0
build==0.10.0
celery==5.3.6
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.1
defusedxml==0.7.1
dill==0.3.6
Django==4.2.1
django-ckeditor==6.5.1
django-compat==1.0.15
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.5
django-guardian==2.4.0
django-js-asset==2.0.0
django-pdb==0.6.2
django-sslserver==0.22
django-storages==1.14.2
django-taggit==4.0.0
django-werkzeug-debugger-runserver==0.3.1
django4-background-tasks==1.2.7
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt==5.3.1
docutils==0.20.1
dodgy==0.2.1
flake8==5.0.4
flake8-polyfill==1.0.2
flower==2.0.1
gitdb==4.0.10
GitPython==3.1.31
httpie==3.2.2
humanize==4.9.0
idna==3.4
isort==5.12.0
kombu==5.3.4
lazy-object-proxy==1.9.0
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.4
mypy==1.3.0
mypy-extensions==1.0.0
mysqlclient==2.1.1
nodeenv==1.8.0
olefile==0.47
packaging==23.1
paramiko==3.4.0
pbr==5.11.1
pep8-naming==0.10.0
Pillow==9.5.0
platformdirs==3.5.1
prometheus-client==0.19.0
prompt-toolkit==3.0.43
prospector==1.10.2
pycodestyle==2.9.1
pycparser==2.21
pydocstyle==6.3.0
pyflakes==2.5.0
Pygments==2.15.1
PyJWT==2.8.0
pylint==2.17.4
pylint-celery==0.3
pylint-django==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.7
PyMySQL==1.1.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyproject_hooks==1.0.0
pyright==1.1.311
pyroma==4.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==5.0.1
requests==2.31.0
requests-toolbelt==1.0.0
requirements-detector==1.2.2
rich==13.4.1
semver==3.0.0
setoptconf-tmp==0.3.1
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
soupsieve==2.4.1
sqlparse==0.4.4
stevedore==5.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
tornado==6.4
trove-classifiers==2023.5.24
typing_extensions==4.6.2
tzdata==2023.4
urllib3==2.0.2
uWSGI==2.0.23
vine==5.1.0
vulture==2.7
wcwidth==0.2.12
Werkzeug==2.3.6
wrapt==1.15.0
